refactor: restructure project feature and enhance project creation logic
- Introduced a new ProjectController to manage project-related HTTP requests.
- Created CreateProjectDto for structured project creation data validation.
- Updated project module to reflect new controller and DTO paths.
- Refactored project service to utilize the new DTO for improved type safety and validation.
- Enhanced overall project structure to follow a feature-first approach, aligning with NestJS best practices.
