+++ apps/server/src/features/project/project.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { PrismaModule } from 'prisma/prisma.module';
 import { ProjectService } from './services/project.service';
 import { PrismaProjectRepository } from './repositories/prisma.project.repository';
-import { ProjectController } from './project.controller';
+import { ProjectController } from './controllers/project.controller';
 import { PROJECT_REPOSITORY } from './repositories/project.repository.interface';
 import { TechStackModule } from '../tech-stack/tech-stack.module';
 import { CategoryModule } from '../category/category.module';
